Change volume DB type to uint16_t

+++ stereo_microphone_i2s/inc/microphone_settings.h
@@ -13,7 +13,7 @@ typedef struct _microphone_settings_t {
   // Current states
   int8_t mute[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX + 1];       // +1 for master channel 0
   int16_t volume[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX + 1];    // +1 for master channel 0
-  int32_t volume_db[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX + 1];    // +1 for master channel 0
+  uint16_t volume_db[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_TX + 1];    // +1 for master channel 0
 
   uint32_t sample_rate;
   uint8_t resolution;
